NFT unstaking and claim added.

+++ nft-staking/programs/nft-staking/src/state/user_account.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 
 #[account]
 pub struct UserAccount {
-    pub points_earned:u32,
+    pub points_earned:i64,
     pub amount_staked: u8,
     pub bump: u8
 }
